requirements.txt
Babel==1.3
coverage==3.7.1
docutils==0.12
Flask==0.10.1
Flask-Testing==0.4.2
Flask-Menu==0.4.0
gevent==1.0.1
gevent-websocket==0.9.3
greenlet==0.4.7
itsdangerous==0.24
Jinja2==2.7.3
jsonschema==2.4.0
MarkupSafe==0.23
nose==1.3.7
ptyprocess==0.4
pytz==2015.4
pyzmq==14.7.0
six==1.9.0
ujson==1.33
Werkzeug==0.9.6
